requirements.txt
pandas==2.2.2
streamlit==1.37.0
streamlit-autorefresh==1.0.1
python-dotenv==1.0.1
